LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-embed-element/embed-network-error.sub-expected.txt

Blocked access to external URL http://nonexistent.localhost/
CONSOLE MESSAGE: Refused to display 'http://localhost:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.
Blocked access to external URL http://nonexistent.localhost/
CONSOLE MESSAGE: Refused to display 'http://localhost:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.


PASS new embed: nonexistent host
PASS new embed: X-Frame-Options prevents embedding
PASS navigating an existing embed: nonexistent host
PASS navigating an existing embed: X-Frame-Options prevents embedding
